feat: Implemented the dice roll feature as a seperate widget
