+++ frontend/lib/pages/home/home_page.dart
@@ -5,10 +5,8 @@ import 'package:nibbles/service/cuisine/cuisine_service.dart';
 import 'package:nibbles/service/profile/restaurant_dto.dart';
 import 'package:nibbles/service/restaurant/restaurant_service.dart';
 import 'package:nibbles/theme/app_theme.dart';
+import 'package:nibbles/pages/recipes/widgets/dice_widget.dart';
 import 'dart:math';
-
-// Import the dice roll widget (assuming it's in the same package)
-// import 'dice_roll_widget.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -331,220 +329,6 @@ class _HomePageState extends State<HomePage> {
           ),
           _buildActiveFiltersChip(),
         ],
-      ),
-    );
-  }
-}
-
-// Dice Roll Widget and Modal (inline for this example)
-class DiceRollWidget extends StatefulWidget {
-  final VoidCallback onPressed;
-  final bool isEnabled;
-
-  const DiceRollWidget({
-    super.key,
-    required this.onPressed,
-    this.isEnabled = true,
-  });
-
-  @override
-  State<DiceRollWidget> createState() => _DiceRollWidgetState();
-}
-
-class _DiceRollWidgetState extends State<DiceRollWidget>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _animationController;
-  late Animation<double> _rotationAnimation;
-  late Animation<double> _scaleAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 800),
-      vsync: this,
-    );
-
-    _rotationAnimation = Tween<double>(begin: 0.0, end: 2.0).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.easeInOut),
-    );
-
-    _scaleAnimation = Tween<double>(begin: 1.0, end: 1.2).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.elasticOut),
-    );
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-
-  void _handlePress() {
-    if (!widget.isEnabled) return;
-
-    _animationController.forward().then((_) {
-      _animationController.reset();
-      widget.onPressed();
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-
-    return AnimatedBuilder(
-      animation: _animationController,
-      builder: (context, child) {
-        return Transform.scale(
-          scale: _scaleAnimation.value,
-          child: Transform.rotate(
-            angle: _rotationAnimation.value * pi,
-            child: IconButton(
-              onPressed: widget.isEnabled ? _handlePress : null,
-              icon: Icon(
-                Icons.casino,
-                color:
-                    widget.isEnabled
-                        ? Colors.white
-                        : colorScheme.onSurfaceVariant.withOpacity(0.5),
-                size: 24,
-              ),
-              tooltip: 'Random cuisine selection',
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
-
-class DiceResultModal extends StatefulWidget {
-  final String title;
-  final String subtitle;
-  final IconData icon;
-  final VoidCallback onClose;
-
-  const DiceResultModal({
-    super.key,
-    required this.title,
-    required this.subtitle,
-    required this.icon,
-    required this.onClose,
-  });
-
-  @override
-  State<DiceResultModal> createState() => _DiceResultModalState();
-}
-
-class _DiceResultModalState extends State<DiceResultModal>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _animationController;
-  late Animation<double> _scaleAnimation;
-  late Animation<double> _opacityAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 600),
-      vsync: this,
-    );
-
-    _scaleAnimation = Tween<double>(begin: 0.5, end: 1.0).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.elasticOut),
-    );
-
-    _opacityAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.easeIn),
-    );
-
-    _animationController.forward();
-
-    // Auto close after 2.5 seconds
-    Future.delayed(const Duration(milliseconds: 2500), () {
-      if (mounted) {
-        _close();
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    super.dispose();
-  }
-
-  void _close() {
-    _animationController.reverse().then((_) {
-      if (mounted) {
-        widget.onClose();
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final textTheme = Theme.of(context).textTheme;
-
-    return Material(
-      color: Colors.black54,
-      child: Center(
-        child: AnimatedBuilder(
-          animation: _animationController,
-          builder: (context, child) {
-            return Opacity(
-              opacity: _opacityAnimation.value,
-              child: Transform.scale(
-                scale: _scaleAnimation.value,
-                child: Container(
-                  margin: const EdgeInsets.all(32),
-                  padding: const EdgeInsets.all(24),
-                  decoration: BoxDecoration(
-                    color: colorScheme.surface,
-                    borderRadius: BorderRadius.circular(16),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black.withOpacity(0.3),
-                        blurRadius: 10,
-                        offset: const Offset(0, 5),
-                      ),
-                    ],
-                  ),
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Icon(widget.icon, size: 48, color: colorScheme.primary),
-                      const SizedBox(height: 16),
-                      Text(
-                        widget.title,
-                        style: textTheme.headlineSmall?.copyWith(
-                          color: colorScheme.onSurface,
-                          fontWeight: FontWeight.bold,
-                        ),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(height: 8),
-                      Text(
-                        widget.subtitle,
-                        style: textTheme.bodyMedium?.copyWith(
-                          color: colorScheme.onSurfaceVariant,
-                        ),
-                        textAlign: TextAlign.center,
-                      ),
-                      const SizedBox(height: 16),
-                      TextButton(
-                        onPressed: _close,
-                        child: const Text('Got it!'),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            );
-          },
-        ),
       ),
     );
   }
