+++ frontend/lib/pages/recipes/recipes_page.dart
@@ -107,6 +107,28 @@ class _RecipesPageState extends State<RecipesPage> {
   }
 
   Future<void> _generateRecipes() async {
+    // Check if at least one appliance is selected
+    if (selectedAppliances.isEmpty) {
+      showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: const Text('Missing Kitchen Appliances'),
+            content: const Text(
+              'Please select at least one kitchen appliance to generate recipes.',
+            ),
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.of(context).pop(),
+                child: const Text('OK'),
+              ),
+            ],
+          );
+        },
+      );
+      return;
+    }
+
     // Show loading dialog
     showDialog(
       context: context,
@@ -407,9 +429,19 @@ class _RecipesPageState extends State<RecipesPage> {
                         child: SizedBox(
                           width: double.infinity,
                           child: ElevatedButton.icon(
-                            onPressed: _generateRecipes,
+                            onPressed:
+                                selectedAppliances.isNotEmpty
+                                    ? _generateRecipes
+                                    : null,
                             icon: Icon(Icons.restaurant_menu),
                             label: Text('Generate Recipes'),
+                            style:
+                                selectedAppliances.isEmpty
+                                    ? ElevatedButton.styleFrom(
+                                      backgroundColor: Colors.grey.shade300,
+                                      foregroundColor: Colors.grey.shade600,
+                                    )
+                                    : null,
                           ),
                         ),
                       ),
